+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+
+
+## [0.3.6] - 2025-02-10
 ### Fixed
 - Don't reuse connections (leaves PHP-FPM connections hanging)
 - force enabled buffer in Shared Traceable client response
@@ -61,7 +65,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.0] - 2024-03-27
 - Initial release
 
-[Unreleased]: https://github.com/efabrica-team/http-client/compare/0.3.5...master
+[Unreleased]: https://github.com/efabrica-team/http-client/compare/0.3.6...master
+[0.3.6]: https://github.com/efabrica-team/http-client/compare/0.3.5...0.3.6
 [0.3.5]: https://github.com/efabrica-team/http-client/compare/0.3.3...0.3.5
 [0.3.3]: https://github.com/efabrica-team/http-client/compare/0.3.2...0.3.3
 [0.3.2]: https://github.com/efabrica-team/http-client/compare/0.3.1...0.3.2